Improve merging code in tests/__init__.py. Move dictionary merging code into AATA tests script.
